requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
opencv-python==4.9.0.80
torch>=2.0.0
torchvision>=0.15.0
numpy==1.26.4
shapely==2.0.3
networkx==3.2.1
google-generativeai>=0.8.3
pytesseract==0.3.10
pillow==10.2.0
python-multipart==0.0.9
scipy==1.12.0
stellar-sdk>=9.0.0
aiofiles==23.2.1
python-dotenv==1.0.1
segmentation-models-pytorch==0.3.3
pulp==2.8.0
